fix_update

+++ apple-management/src/main/java/com/example/applemanagement/repository/employee/IPositionRepository.java
@@ -3,5 +3,5 @@ package com.example.applemanagement.repository.employee;
 import com.example.applemanagement.model.employee.Position;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IPositionRepository extends JpaRepository<Position, Integer> {
+public interface IPositionRepository extends JpaRepository<Position, Long> {
 }
